+++ src/components/popover/index.js
@@ -121,7 +121,7 @@ class Popover {
 
 window.Popover = Popover;
 
-function initPopover() {
+export function initPopovers() {
     document.querySelectorAll('[data-popover-target]').forEach(triggerEl => {
         const targetEl = document.getElementById(triggerEl.getAttribute('data-popover-target'))
         const triggerType = triggerEl.getAttribute('data-popover-trigger');
@@ -135,18 +135,5 @@ function initPopover() {
         })
     })
 }
-
-const windowEventListeners = ['load'];
-
-// add "turbo:load" event listener if Turbo is enabled
-if (typeof Rails !== 'undefined' && Rails.application.config.action_controller.use_turbo_stream) {
-    windowEventListeners.pop(); // remove "load" event
-    windowEventListeners.push("turbo:load");
-}
-
-// init collapse on load
-windowEventListeners.forEach((event) => {
-    window.addEventListener(event, initPopover);
-});
 
 export default Popover