+++ src/components/tabs/index.js
@@ -78,7 +78,7 @@ class Tabs {
 
 window.Tabs = Tabs;
 
-function initTabs() {
+export function initTabs() {
     document.querySelectorAll('[data-tabs-toggle]').forEach(triggerEl => {
 
         const tabElements = []
@@ -101,18 +101,5 @@ function initTabs() {
         })
     })
 }
-
-const windowEventListeners = ['load'];
-
-// add "turbo:load" event listener if Turbo is enabled
-if (typeof Rails !== 'undefined' && Rails.application.config.action_controller.use_turbo_stream) {
-    windowEventListeners.pop(); // remove "load" event
-    windowEventListeners.push("turbo:load");
-}
-
-// init collapse on load
-windowEventListeners.forEach((event) => {
-    window.addEventListener(event, initTabs);
-});
 
 export default Tabs