+++ webpack.config.js
@@ -8,6 +8,7 @@ module.exports = {
   entry: {
     datepicker: path.resolve(__dirname, 'src/plugins/datepicker.js'),
     flowbite: path.resolve(__dirname, 'src/flowbite.js'),
+    'flowbite.turbo': path.resolve(__dirname, 'src/flowbite.turbo.js'),
     docs: path.resolve(__dirname, 'src/docs.js')
   },
   output: {